move codet types specific to goto-programs into a separate header file
The codet types are used both at the source level (C/C++/Java) and in goto-programs,
which is confusing.

This commit moves those types that are exclusive to goto-program
instructions into a separate header file.  It is envisaged that this header
file moves to goto-programs/ eventually.
