+++ jbmc/src/java_bytecode/code_with_references.cpp
@@ -8,7 +8,9 @@ Author: Romain Brenguier, romain.brenguier@diffblue.com
 
 #include "code_with_references.h"
 #include "java_types.h"
+
 #include <util/arith_tools.h>
+#include <util/goto_instruction_code.h>
 
 codet allocate_array(
   const exprt &expr,
