+++ jbmc/src/java_bytecode/lift_clinit_calls.cpp
@@ -13,6 +13,7 @@ Author: Diffblue Ltd.
 #include <java_bytecode/java_static_initializers.h>
 
 #include <util/expr_iterator.h>
+#include <util/goto_instruction_code.h>
 
 #include <algorithm>
