+++ jbmc/unit/java-testing-utils/require_goto_statements.h
@@ -9,8 +9,8 @@ Author: Diffblue Ltd.
 /// \file
 /// Utilties for inspecting goto functions
 
+#include <util/goto_instruction_code.h>
 #include <util/optional.h>
-#include <util/std_code.h>
 
 #include <regex>
