+++ src/ansi-c/expr2c_class.h
@@ -18,6 +18,7 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include <util/bitvector_expr.h>
 #include <util/byte_operators.h>
+#include <util/goto_instruction_code.h>
 #include <util/mathematical_expr.h>
 #include <util/std_code.h>
