+++ src/cpp/cpp_typecheck_constructor.cpp
@@ -13,6 +13,7 @@ Author: Daniel Kroening, kroening@cs.cmu.edu
 
 #include <util/arith_tools.h>
 #include <util/c_types.h>
+#include <util/goto_instruction_code.h>
 #include <util/pointer_expr.h>
 #include <util/std_code.h>
