+++ src/cpp/cpp_typecheck_initializer.cpp
@@ -14,6 +14,7 @@ Author: Daniel Kroening, kroening@cs.cmu.edu
 #include <util/arith_tools.h>
 #include <util/c_types.h>
 #include <util/expr_initializer.h>
+#include <util/goto_instruction_code.h>
 #include <util/pointer_expr.h>
 #include <util/pointer_offset_size.h>
