+++ src/goto-instrument/object_id.cpp
@@ -11,8 +11,8 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include "object_id.h"
 
+#include <util/goto_instruction_code.h>
 #include <util/pointer_expr.h>
-#include <util/std_code.h>
 
 enum class get_modet { LHS_R, LHS_W, READ };
