+++ src/goto-programs/destructor.cpp
@@ -11,8 +11,8 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include "destructor.h"
 
+#include <util/goto_instruction_code.h>
 #include <util/namespace.h>
-#include <util/std_code.h>
 
 code_function_callt get_destructor(
   const namespacet &ns,
