+++ src/jsil/jsil_convert.cpp
@@ -11,6 +11,7 @@ Author: Michael Tautschnig, tautschn@amazon.com
 
 #include "jsil_convert.h"
 
+#include <util/goto_instruction_code.h>
 #include <util/message.h>
 #include <util/symbol_table.h>
