+++ src/jsil/jsil_typecheck.cpp
@@ -12,8 +12,8 @@ Author: Michael Tautschnig, tautschn@amazon.com
 #include "jsil_typecheck.h"
 
 #include <util/bitvector_types.h>
+#include <util/goto_instruction_code.h>
 #include <util/prefix.h>
-#include <util/std_code.h>
 #include <util/symbol_table.h>
 
 #include "expr2jsil.h"
