+++ src/pointer-analysis/value_set_fi.cpp
@@ -15,6 +15,8 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include <util/arith_tools.h>
 #include <util/byte_operators.h>
+#include <util/c_types.h>
+#include <util/goto_instruction_code.h>
 #include <util/namespace.h>
 #include <util/pointer_expr.h>
 #include <util/prefix.h>
@@ -23,7 +25,6 @@ Author: Daniel Kroening, kroening@kroening.com
 #include <util/symbol.h>
 
 #include <langapi/language_util.h>
-#include <util/c_types.h>
 
 const value_set_fit::object_map_dt value_set_fit::object_map_dt::blank{};
