+++ src/util/std_code.cpp
@@ -13,6 +13,7 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include "arith_tools.h"
 #include "c_types.h"
+#include "goto_instruction_code.h"
 #include "pointer_expr.h"
 #include "string_constant.h"
